refactor: extract IngestionLeaseManager from IngestionService
Moves lease/run lifecycle methods (TryAcquireIngestionRunAsync,
ReleaseIngestionLeaseAsync, FinalizeAbortedRunIfPendingAsync,
UpdateRuntimeStateAsync, UpdateIngestionRunStatusAsync,
UpdateActiveRunStatusAsync, CompleteIngestionRunAsync,
CleanupExpiredIngestionArtifactsAsync, ThrowIfAbortRequestedAsync)
plus ClearStagedDataForRunAsync into a new IngestionLeaseManager class.
IngestionService now delegates all lease/run lifecycle calls to the
injected collaborator. Registers IngestionLeaseManager in DI and adds
two TDD tests covering lease acquisition.
